For #26706 - Change TopSite.containsQueryParameters as Uri extension function
