+++ app/src/main/java/org/mozilla/fenix/ext/TopSite.kt
@@ -4,7 +4,6 @@
 
 package org.mozilla.fenix.ext
 
-import android.net.Uri
 import mozilla.components.feature.top.sites.TopSite
 import org.mozilla.fenix.settings.SupportUtils
 
@@ -34,33 +33,5 @@ fun List<TopSite>.sort(): List<TopSite> {
         result.removeAt(defaultGoogleTopSiteIndex)
         result.add(0, this[defaultGoogleTopSiteIndex])
         result
-    }
-}
-
-/**
- * Returns true if the url contains any query parameters specified by the [searchParameters].
- *
- * @param searchParameters [String] of the following forms:
- * - "" (empty) - Don't search for any params
- * - "key" - Search param named "key" with any or no value
- * - "key=" - Search param named "key" with no value
- * - "key=value" - Search param named "key" with value "value"
- */
-fun TopSite.containsQueryParameters(searchParameters: String): Boolean {
-    if (searchParameters.isBlank()) {
-        return false
-    }
-    val params = searchParameters.split("=")
-    val uri = Uri.parse(url)
-    return when (params.size) {
-        1 -> {
-            uri.queryParameterNames.contains(params.first()) &&
-                uri.getQueryParameter(params.first()).isNullOrBlank()
-        }
-        2 -> {
-            uri.queryParameterNames.contains(params.first()) &&
-                uri.getQueryParameter(params.first()) == params.last()
-        }
-        else -> false
     }
 }
