+++ app/src/main/java/org/mozilla/fenix/utils/Settings.kt
@@ -1112,8 +1112,8 @@ class Settings(private val appContext: Context) : PreferencesHolder {
         default = ""
     )
     /**
-     *  URLs from the user's history that contain this search param will be hidden
-     * from the top sites. The value is a string with one of the following forms:
+     *  URLs from the user's history that contain this search param will be hidden.
+     *  The value is a string with one of the following forms:
      * - "" (empty) - Disable this feature
      * - "key" - Search param named "key" with any or no value
      * - "key=" - Search param named "key" with no value
@@ -1121,7 +1121,7 @@ class Settings(private val appContext: Context) : PreferencesHolder {
      */
     val frecencyFilterQuery by stringPreference(
         appContext.getPreferenceKey(R.string.pref_key_frecency_filter_query),
-        default = "mfadid=adm"
+        default = "mfadid=adm" // Parameter provided by adM
     )
 
     /**
